base: Enable boot/bootloader/radio version checks
This enables the Build.isBuildConsistent checks after tweaking the radio
check to stop failing when the radio hasn't been turned on. The property
used to check the radio version (gsm.baseband.version) is only set once
the radio is turned on and initialized.

The API documentation already claims that this is done.